+++ packages/moonstone/CHANGELOG.md
@@ -9,6 +9,10 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Scroller`, `moonstone/VirtualList.VirtualGridList`, and `moonstone/VirtualList.VirtualList` to show overscroll effects only by wheel input
 - `moonstone/Icon` and `moonstone/IconButton` to require `children`
 
+### Fixed
+
+- `moonstone/Scrollable` to calculate and `scrollTo()` when updated by resizing
+
 ## [2.1.0] - 2018-08-20
 
 ### Added